Now ignoring local workbench/ dir
I wanted to preserve some "personal" files related to the repo I'm
working on, but didn't want to add a reference to it in the shared
.gitignore file.

My new approach:  globally ignore workbench/, and use that inside the
repos I'm working on.  I haven't heard of others doing this, but I think
it might be kinda cool?